requirements.txt
uvicorn[standard]
fastapi
pydantic
pydantic-settings
motor
google-auth
pyjwt

